Refact: Removed Google id old code, new database population data for tests

+++ application.py
@@ -43,6 +43,13 @@ def showLogin():
     # return "The current session state is %s" % login_session['state']
     return render_template('login.html', STATE=state)
 
+@app.route('/logout')
+def aLogout():
+    if login_session['provider'] == 'facebook':
+        fbdisconnect()
+        login_session.clear()
+        return redirect(url_for('showHome'))    
+
 
 @app.route('/fbconnect', methods=['POST'])
 def fbconnect():
@@ -122,125 +129,10 @@ def fbdisconnect():
     url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (
         facebook_id, access_token)
     h = httplib2.Http()
-    result = h.request(url, 'DELETE')[1]
+    result = h.request(url, 'GET')[0]
     return "you have been logged out"
 
 
-# @app.route('/gconnect', methods=['POST'])
-# def gconnect():
-#     # Validate state token
-#     if request.args.get('state') != login_session['state']:
-#         response = make_response(json.dumps('Invalid state parameter.'), 401)
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-#     # Obtain authorization code
-#     code = request.data
-
-#     try:
-#         # Upgrade the authorization code into a credentials object
-#         oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
-#         oauth_flow.redirect_uri = 'postmessage'
-#         credentials = oauth_flow.step2_exchange(code)
-#     except FlowExchangeError:
-#         response = make_response(
-#             json.dumps('Failed to upgrade the authorization code.'), 401)
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-
-#     # Check that the access token is valid.
-#     access_token = credentials.access_token
-#     url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
-#            % access_token)
-#     h = httplib2.Http()
-#     result = json.loads(h.request(url, 'GET')[1])
-#     # If there was an error in the access token info, abort.
-#     if result.get('error') is not None:
-#         response = make_response(json.dumps(result.get('error')), 500)
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-
-#     # Verify that the access token is used for the intended user.
-#     gplus_id = credentials.id_token['sub']
-#     if result['user_id'] != gplus_id:
-#         response = make_response(
-#             json.dumps("Token's user ID doesn't match given user ID."), 401)
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-
-#     # Verify that the access token is valid for this app.
-#     if result['issued_to'] != CLIENT_ID:
-#         response = make_response(
-#             json.dumps("Token's client ID does not match app's."), 401)
-#         print("Token's client ID does not match app's.")
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-
-#     stored_access_token = login_session.get('access_token')
-#     stored_gplus_id = login_session.get('gplus_id')
-#     if stored_access_token is not None and gplus_id == stored_gplus_id:
-#         response = make_response(json.dumps('Current user is already connected.'),
-#                                  200)
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-
-#     # Store the access token in the session for later use.
-#     login_session['access_token'] = credentials.access_token
-#     login_session['gplus_id'] = gplus_id
-
-#     # Get user info
-#     userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
-#     params = {'access_token': credentials.access_token, 'alt': 'json'}
-#     answer = requests.get(userinfo_url, params=params)
-
-#     data = answer.json()
-
-#     login_session['username'] = data['name']
-#     login_session['picture'] = data['picture']
-#     login_session['email'] = data['email']
-#     # ADD PROVIDER TO LOGIN SESSION
-#     login_session['provider'] = 'google'
-
-#     # see if user exists, if it doesn't make a new one
-#     user_id = getUserID(data["email"])
-#     if not user_id:
-#         user_id = createUser(login_session)
-#     login_session['user_id'] = user_id
-
-#     output = ''
-#     output += '<h1>Welcome, '
-#     output += login_session['username']
-#     output += '!</h1>'
-#     output += '<img src="'
-#     output += login_session['picture']
-#     output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
-#     flash("you are now logged in as %s" % login_session['username'])
-#     print("done!")
-#     return output
-
-# # DISCONNECT - Revoke a current user's token and reset their login_session
-
-
-# @app.route('/gdisconnect')
-# def gdisconnect():
-#     # Only disconnect a connected user.
-#     access_token = login_session.get('access_token')
-#     if access_token is None:
-#         response = make_response(
-#             json.dumps('Current user not connected.'), 401)
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-#     url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
-#     h = httplib2.Http()
-#     result = h.request(url, 'GET')[0]
-#     if result['status'] == '200':
-#         response = make_response(json.dumps('Successfully disconnected.'), 200)
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
-#     else:
-#         response = make_response(json.dumps(
-#             'Failed to revoke token for given user.', 400))
-#         response.headers['Content-Type'] = 'application/json'
-#         return response
 
 
 # User Helper Functions
@@ -265,7 +157,6 @@ def getUserID(email):
         return user.id
     except:
         return None
-
 
 # Flask Routes
 # all catalog - JSON
@@ -312,7 +203,7 @@ def ItemsJSON(category_id, item_id):
 @app.route('/')
 def showHome():
     categories = session.query(Category).all()
-    items = session.query(Item).order_by(Item.id.desc())
+    items = session.query(Item).order_by(Item.id.desc())[0:12]
     return render_template('showHome.html', categories=categories, items=items)
 
 
@@ -323,7 +214,6 @@ def showCatalogCategory(category_id):
     categories = session.query(Category).all()
     try:
         category = session.query(Category).filter_by(id=category_id).one()
-
     except NoResultFound:
         flash("Error: The category with id '%s' does not exist." %
               category_id, "error")
@@ -331,21 +221,8 @@ def showCatalogCategory(category_id):
 
     items = session.query(Item).filter_by(cat_id=category_id).all()
     return render_template('showCatalogCategory.html', categories=categories,
-                           sCategory=category, items=items)
+                           sCategory=category, items=items, login_session=login_session)
 
-
-# Catalog item web page
-@app.route('/categories/<int:category_id>/items/<int:item_id>/')
-def showCatalogItem(category_id, item_id):
-    categories = session.query(Category).all()
-    try:
-        item = session.query(Item).filter_by(id=item_id).one()
-        category = session.query(Category).filter_by(id=item.cat_id).one()
-    except NoResultFound:
-        flash("Error: The Item with id '%s' does not exist." % item_id, "error")
-        return redirect(url_for('showHome'))
-    return render_template('showCatalogItem.html', category=category,
-                           item=item, categories=categories)
 
 # create a new Catalog Category
 @app.route('/admin/categories/new/', methods=['GET', 'POST'])
@@ -357,7 +234,7 @@ def newCategory():
         flash("New Category Created!", "success")
         return redirect(url_for('newCategory'))
     else:
-        return render_template('newCategory.html')
+        return render_template('newCategory.html', login_session=login_session)
 
 
 # Edit a Catalog Category
@@ -371,14 +248,12 @@ def editCategory(category_id):
               category_id, "error")
         return redirect(url_for('showHome'))
     if request.method == 'POST':
-        if request.form['name']:
-            editedCategory.name = request.form['name']
         session.add(editedCategory)
         session.commit()
         flash("Category updated!", "success")
         return redirect(url_for('editCategory', category_id=editedCategory.id))
     else:
-        return render_template('editCategory.html', category=editedCategory)
+        return render_template('editCategory.html', category=editedCategory, login_session=login_session)
 
 
 # Delete a Category
@@ -397,23 +272,43 @@ def deleteCategory(category_id):
         flash("Category deleted!", "success")
         return redirect(url_for('showHome'))
     else:
-        return render_template('deleteCategory.html', category=categoryToDelete)
+        return render_template('deleteCategory.html', category=categoryToDelete, login_session=login_session)
+
+
+# Catalog item web page
+@app.route('/categories/<int:category_id>/items/<int:item_id>/')
+def showCatalogItem(category_id, item_id):
+    categories = session.query(Category).all()
+    try:
+        item = session.query(Item).filter_by(id=item_id).one()
+        category = session.query(Category).filter_by(id=item.cat_id).one()
+    except NoResultFound:
+        flash("Error: The Item with id '%s' does not exist." % item_id, "error")
+        return redirect(url_for('showHome'))
+    return render_template('showCatalogItem.html', category=category,
+                           item=item, categories=categories, login_session=login_session)
+
 
 # Create Catalog Item
 @app.route('/items/new/', methods=['GET', 'POST'])
 @app.route('/categories/<int:category_id>/items/new/', methods=['GET', 'POST'])
 def newItem(category_id=0):
     categories = session.query(Category).all()
-
+    if 'username' not in login_session:
+        flash("Please, login to your account to proceed!", "error")
+        if category_id == 0 :
+            return redirect(url_for('showHome'))
+        else:
+            return redirect(url_for('showCatalogCategory', category_id=category_id))
     if request.method == 'POST':
         newItem = Item(title=request.form['title'], description=request.form['description'],
-                       cat_id=request.form['categoryId'], user_id=1)
+                       cat_id=request.form['categoryId'], user_id=login_session['user_id'])
         session.add(newItem)
         session.commit()
         flash("Item created!", "success")
-        return redirect(url_for('newItem', category_id=newItem.cat_id))
+        return redirect(url_for('showCatalogItem', category_id=newItem.cat_id, item_id=newItem.id))
     else:
-        return render_template('newItem.html', categories=categories, category_id=category_id)
+        return render_template('newItem.html', categories=categories, category_id=category_id, login_session=login_session)
 
 
 # Edit a Catalog Item
@@ -424,7 +319,13 @@ def editItem(category_id, item_id):
         editedItem = session.query(Item).filter_by(id=item_id).one()
     except NoResultFound:
         flash("Error: The Item with id '%s' does not exist." % item_id, "error")
-        return redirect(url_for('showHome'))
+        return redirect(url_for('showCatalogCategory', category_id=category_id))
+    if 'username' not in login_session:
+        flash("Please, login to your account to proceed!", "error")
+        return redirect(url_for('showCatalogItem', category_id=editedItem.cat_id, item_id=item_id))
+    if editedItem.user_id != login_session['user_id']:
+        flash("Only the Owner can edit this Item", "error")
+        return redirect(url_for('showCatalogItem', category_id=editedItem.cat_id, item_id=item_id))
     if request.method == 'POST':
         if request.form['title']:
             editedItem.title = request.form['title']
@@ -436,12 +337,12 @@ def editItem(category_id, item_id):
         session.add(editedItem)
         session.commit()
         flash("Item updated!", "success")
-        return redirect(url_for('editItem', category_id=editedItem.cat_id, item_id=editedItem.id))
+        return redirect(url_for('showCatalogItem', category_id=editedItem.cat_id, item_id=editedItem.id))
     else:
         if editedItem.cat_id != category_id:
             return redirect(url_for('editItem', category_id=editedItem.cat_id, item_id=editedItem.id))
         else:
-            return render_template('editItem.html', item=editedItem, categories=categories)
+            return render_template('editItem.html', item=editedItem, categories=categories, login_session=login_session)
 
 
 # Delete a Category item
@@ -451,17 +352,23 @@ def deleteItem(category_id, item_id):
         itemToDelete = session.query(Item).filter_by(id=item_id).one()
     except NoResultFound:
         flash("Error: The Item with id '%s' does not exist." % item_id, "error")
-        return redirect(url_for('showHome'))
+        return redirect(url_for('showCatalogCategory', category_id=category_id))
+    if 'username' not in login_session:
+        flash("Please, login to your account to proceed!", "error")
+        return redirect(url_for('showCatalogItem', category_id=itemToDelete.cat_id, item_id=item_id))
+    if itemToDelete.user_id != login_session['user_id']:
+        flash("Only the Owner can delete this Item", "error")
+        return redirect(url_for('showCatalogItem', category_id=itemToDelete.cat_id, item_id=item_id))
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
         flash("Item deleted!", "success")
-        return redirect(url_for('showHome'))
+        return redirect(url_for('showCatalogCategory', category_id=category_id))
     else:
         if itemToDelete.cat_id != category_id:
             return redirect(url_for('deleteItem', category_id=itemToDelete.cat_id, item_id=itemToDelete.id))
         else:
-            return render_template('deleteItem.html', item=itemToDelete)
+            return render_template('deleteItem.html', item=itemToDelete, login_session=login_session)
 
 
 if __name__ == '__main__':
